+++ dqmj1_randomizer/state.py
@@ -1,8 +1,37 @@
 import pathlib
+import re
 from dataclasses import dataclass, field
 from typing import Optional
 
 from dqmj1_randomizer.randomize.regions import Region
+
+
+@dataclass(frozen=True)
+class FullyRandomMonsterShuffle:
+    pass
+
+
+@dataclass(frozen=True)
+class BiasedByStatTotalMonsterShuffle:
+    leniency: int
+
+
+MonsterRandomizationPolicyDefinition = (
+    FullyRandomMonsterShuffle | BiasedByStatTotalMonsterShuffle
+)
+
+
+def parse_monster_randomization_policy_definition(
+    definition_str: str,
+) -> Optional[MonsterRandomizationPolicyDefinition]:
+    if definition_str == "Fully Random":
+        return FullyRandomMonsterShuffle()
+
+    match = re.search(r"([0-9]+)", definition_str)
+    if match is not None:
+        return BiasedByStatTotalMonsterShuffle(int(match.groups()[0]))
+
+    return None
 
 
 @dataclass
@@ -12,6 +41,10 @@ class Monsters:
     transfer_boss_item_drops: Optional[bool] = None
     include_starters: Optional[bool] = None
     include_gift_monsters: Optional[bool] = None
+    swap_scout_chance: Optional[bool] = None
+    swap_experience_drops: Optional[bool] = None
+    swap_gold_drops: Optional[bool] = None
+    randomization_policy: Optional[MonsterRandomizationPolicyDefinition] = None
 
 
 @dataclass